RAG/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-ollama
pydantic
python-dotenv
httpx
requests

# document loader & Embedding
pdfplumber
unstructured[pdf]
pymupdf
statistics
python-docx
sqlalchemy
psycopg2-binary
langchain_huggingface
langchain_postgres
sentence-transformers